+++ 20_Spring_WebFlux/src/main/java/ru/otus/spring/webflux/controller/CommentController.java
@@ -1,11 +1,11 @@
-package ru.otus.pk.spring.controller;
+package ru.otus.spring.webflux.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import ru.otus.pk.spring.domain.Comment;
-import ru.otus.pk.spring.repository.CommentRepository;
+import ru.otus.spring.webflux.domain.Comment;
+import ru.otus.spring.webflux.repository.CommentRepository;
 
 import static org.springframework.http.HttpStatus.CREATED;
