+++ 20_Spring_WebFlux/src/main/java/ru/otus/spring/webflux/controller/GenreController.java
@@ -1,20 +1,20 @@
-package ru.otus.pk.spring.controller;
+package ru.otus.spring.webflux.controller;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 import reactor.core.publisher.Flux;
-import ru.otus.pk.spring.domain.Author;
-import ru.otus.pk.spring.repository.BookRepository;
+import ru.otus.spring.webflux.domain.Genre;
+import ru.otus.spring.webflux.repository.BookRepository;
 
 @RequiredArgsConstructor
 @RestController
-public class AuthorController {
+public class GenreController {
 
     private final BookRepository repository;
 
-    @GetMapping("/authors")
-    public Flux<Author> finAll() {
-        return repository.findAllAuthors();
+    @GetMapping("/genres")
+    public Flux<Genre> finAll() {
+        return repository.findAllGenres();
     }
 }
