+++ 20_Spring_WebFlux/src/main/java/ru/otus/spring/webflux/repository/CommentRepository.java
@@ -1,9 +1,9 @@
-package ru.otus.pk.spring.repository;
+package ru.otus.spring.webflux.repository;
 
 import org.springframework.data.mongodb.repository.ReactiveMongoRepository;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
-import ru.otus.pk.spring.domain.Comment;
+import ru.otus.spring.webflux.domain.Comment;
 
 public interface CommentRepository extends ReactiveMongoRepository<Comment, String> {
